# Add validation dataset loss to distributed SFT recipies