#### Context
What is the purpose of this PR? Is it to
- [x] add a new feature
- [ ] fix a bug
- [ ] update tests and/or documentation
- [ ] other (please add here)

Addresses #1042 / part of the #883 for distributed recipies

#### Changelog
What are the changes made in this PR?
* distributed SFT recipeis (full_finetune_distributed and lora_finetune_distributed) support computing loss on validation dataset (no early stopping)
* example of configuring validation dataset in [recipes/configs/llama3_2/1B_qlora_validation.yaml](recipes/configs/llama3_2/1B_qlora_validation.yaml)

New [configuration](https://github.com/pytorch/torchtune/pull/2238#pullrequestreview-2553490828):
```
dataset_validation: null
run_val_every_n_steps: null
max_validation_batches: -1
```

#### Test plan

As [suggested](https://github.com/pytorch/torchtune/pull/2238/files#r1917299528), I run with  compile + opt_in_bwd + activation ckpt + activation offloading on 2xH100 using [mhenrichsen/alpaca_2k_test](https://huggingface.co/datasets/mhenrichsen/alpaca_2k_test) as a validation dataset.

#### LoRA Distributed

```
pip install -U git+https://github.com/bzz/torchtune.git@validate_distributed

# with validation
tune run --nnodes 1 --nproc_per_node 2 lora_finetune_distributed --config 1B_qlora_validation.yaml \
	compile=True \
	enable_activation_checkpointing=True \
	enable_activation_offloading=True \
	optimizer_in_bwd=True \
	tokenizer.max_seq_len=4096 \
	gradient_accumulation_steps=1 \
	epochs=1 \
	batch_size=16 \
	metric_logger._component_=torchtune.training.metric_logging.WandBLogger \
	output_dir=/tmp/torchtune/llama3_2_1B/qlora_val
```

Without validation
```
tune run --nnodes 1 --nproc_per_node 2 lora_finetune_distributed --config 1B_qlora_validation.yaml \
	run_val_every_n_steps=None \
	compile=True \
	enable_activation_checkpointing=True \
	enable_activation_offloading=True \
	optimizer_in_bwd=True \
	tokenizer.max_seq_len=4096 \
	gradient_accumulation_steps=1 \
	epochs=1 \
	batch_size=16 \
	metric_logger._component_=torchtune.training.metric_logging.WandBLogger \
	output_dir=/tmp/torchtune/llama3_2_1B/qlora
```

<img width="811" alt="Screenshot 2025-03-06 at 08 05 19" src="https://github.com/user-attachments/assets/bb87a934-bb4f-4fbd-994d-e96ad27f789c" />



<details>
<description>llama3_2/1B_qlora_validation.yaml</description>

```
diff recipes/configs/llama3_2/1B_qlora_validation.yaml recipes/configs/llama3_2/1B_lora.yaml

20c20
< output_dir: /tmp/torchtune/llama3_2_1B/qlora # /tmp may be deleted by your system. Change it to your preference.
---
> output_dir: /tmp/torchtune/llama3_2_1B/lora # /tmp may be deleted by your system. Change it to your preference.
30c30
<   _component_: torchtune.models.llama3_2.qlora_llama3_2_1b
---
>   _component_: torchtune.models.llama3_2.lora_llama3_2_1b
57,65d56
< # Validation
< dataset_validation:
<   _component_: torchtune.datasets.alpaca_dataset
<   source: mhenrichsen/alpaca_2k_test
<   split: train
<   packed: False  # True increases speed
< run_val_every_n_steps: 100
< max_validation_batches: -1
<
82c73
< gradient_accumulation_steps: 1  # Use to increase effective batch size
---
> gradient_accumulation_steps: 8  # Use to increase effective batch size
84c75
< compile: True  # torch.compile the model + loss, True increases speed + decreases memory
---
> compile: False  # torch.compile the model + loss, True increases speed + decreases memory
88c79
<   _component_: torchtune.training.metric_logging.WandBLogger
---
>   _component_: torchtune.training.metric_logging.DiskLogger
```
</details>

The difference in throughput is due to both runs using the same 2 GPUs.


#### Full Distributed

```
tune cp llama3_2/1B_full 1B_full_validation

# Validation
dataset_validation:
  _component_: torchtune.datasets.alpaca_dataset
  source: mhenrichsen/alpaca_2k_test
  split: train
run_val_every_n_steps: 100
max_validation_batches: -1

metric_logger:
	_component_: torchtune.training.metric_logging.WandBLogger
```

Validation
```
tune run --nnodes 1 --nproc_per_node 2 \
    full_finetune_distributed --config 1B_full_validation.yaml \
	compile=True \
	enable_activation_checkpointing=True \
	enable_activation_offloading=True \
	optimizer_in_bwd=True \
	tokenizer.max_seq_len=4096 \
	gradient_accumulation_steps=1 \
	epochs=1 \
	batch_size=16 \
	output_dir=/tmp/torchtune/llama3_2_1B/full_dist_val
```

And without
```
tune run --nnodes 1 --nproc_per_node 2 \
    full_finetune_distributed --config 1B_full_validation.yaml \
	run_val_every_n_steps=None \
	compile=True \
	enable_activation_checkpointing=True \
	enable_activation_offloading=True \
	optimizer_in_bwd=True \
	tokenizer.max_seq_len=4096 \
	gradient_accumulation_steps=1 \
	epochs=1 \
	batch_size=16 \
	output_dir=/tmp/torchtune/llama3_2_1B/full_dist
```

<img width="807" alt="Screenshot 2025-03-06 at 11 34 48" src="https://github.com/user-attachments/assets/c446bba8-8acf-4648-8529-0bb5ed40babe" />



- [x] run pre-commit hooks and linters (make sure you've first installed via `pre-commit install`)
- [ ] add [unit tests](https://github.com/pytorch/torchtune/tree/main/tests/torchtune) for any new functionality
- [ ] update [docstrings](https://github.com/pytorch/torchtune/tree/main/docs/source) for any new or updated methods or classes
- [ ] run unit tests via `pytest tests`
- [ ] run recipe tests via `pytest tests -m integration_test`
- [x] manually run any new or modified recipes with sufficient proof of correctness
- [x] include relevant commands and any other artifacts in this summary (pastes of loss curves, eval results, etc.)

#### UX

The output is really simplistic and looks like 
```
1|100|Loss: 3.8272886276245117:   6%|████                                                               | 100/1625 [01:05<05:46,  4.41it/s]
INFO:torchtune.utils._logging:Validation loss: 4.8242
1|200|Loss: 1.7956730127334595:  12%|████████▏                                                          | 200/1625 [01:30<05:11,  4.58it/s]
INFO:torchtune.utils._logging:Validation loss: 1.8639
1|300|Loss: 1.4479581117630005:  18%|████████████▎                                                      | 300/1625 [01:52<03:54,  5.66it/s]
INFO:torchtune.utils._logging:Validation loss: 1.7377
1|400|Loss: 1.433119535446167:  25%|████████████████▋                                                   | 400/1625 [02:13<03:51,  5.28it/s]
INFO:torchtune.utils._logging:Validation loss: 1.7247
1|500|Loss: 1.5169874429702759:  31%|████████████████████▌                                              | 500/1625 [02:34<03:13,  5.81it/s]
INFO:torchtune.utils._logging:Validation loss: 1.7714
1|600|Loss: 1.4446262121200562:  37%|████████████████████████▋                                          | 600/1625 [02:55<02:54,  5.89it/s]
INFO:torchtune.utils._logging:Validation loss: 1.7542
1|700|Loss: 1.5118173360824585:  43%|████████████████████████████▊                                      | 700/1625 [03:18<02:54,  5.32it/s]
INFO:torchtune.utils._logging:Validation loss: 1.7290
1|800|Loss: 1.2981687784194946:  49%|████████████████████████████████▉                                  | 800/1625 [03:40<02:09,  6.35it/s]
INFO:torchtune.utils._logging:Validation loss: 1.7309
1|900|Loss: 1.533372163772583:  55%|█████████████████████████████████████▋                              | 900/1625 [04:01<02:22,  5.08it/s]
INFO:torchtune.utils._logging:Validation loss: 1.7264
```

Please if you think it needs to be somehow improved.

- [x] I did not change any public API
- [ ] I have added an example to docs or docstrings