@bzz 2025-04-02
lora

```
        run_val_every_n_steps=100
	batch_size=16 
	batch_size_validation=64 
```

<img width="1170" alt="Screenshot 2025-04-02 at 23 27 33" src="https://github.com/user-attachments/assets/a633703c-9882-4586-a3fb-9ac152c47382" />