@bzz 2025-04-11
I believe all the feedback (but a unit test) should be addressed now. 

@joecummings 

> Overall, I'm curious if we've tried compiling the validation step or this new larger loss step to see any speedups?

I'm not sure I follow - do you suggest to compare training times with validation loop enabled between compile/no compile experiments to see the speedup?